requirements.txt
python-multipart
fastapi
uvicorn
python-dotenv
openai
langchain
faiss-cpu
pypdf
sentence-transformers
streamlit
